# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification for Overhype.me
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /facts:
    get:
      operationId: listFacts
      tags: [facts]
      summary: List facts
      parameters:
        - name: search
          in: query
          required: false
          schema:
            type: string
        - name: hashtag
          in: query
          required: false
          schema:
            type: string
        - name: sort
          in: query
          required: false
          schema:
            type: string
            enum: [top, newest, trending]
            default: top
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 20
            minimum: 1
            maximum: 100
        - name: offset
          in: query
          required: false
          schema:
            type: integer
            default: 0
            minimum: 0
      responses:
        "200":
          description: List of facts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FactListResponse"
    post:
      operationId: createFact
      tags: [facts]
      summary: Submit a new fact
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateFactRequest"
      responses:
        "201":
          description: Fact created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FactDetail"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "409":
          description: Possible duplicate detected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DuplicateConflict"
  /facts/{factId}:
    get:
      operationId: getFact
      tags: [facts]
      summary: Get a single fact with details
      parameters:
        - name: factId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Fact detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FactDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /facts/{factId}/rating:
    post:
      operationId: rateFact
      tags: [ratings]
      summary: Rate a fact (upvote or downvote)
      parameters:
        - name: factId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RateFactRequest"
      responses:
        "200":
          description: Rating updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RatingResult"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /facts/{factId}/comments:
    get:
      operationId: listComments
      tags: [comments]
      summary: List comments for a fact
      parameters:
        - name: factId
          in: path
          required: true
          schema:
            type: integer
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          required: false
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: List of comments
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommentListResponse"
    post:
      operationId: addComment
      tags: [comments]
      summary: Add a comment to a fact
      parameters:
        - name: factId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddCommentRequest"
      responses:
        "201":
          description: Comment created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Comment"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /facts/{factId}/links:
    get:
      operationId: listLinks
      tags: [links]
      summary: List external links for a fact
      parameters:
        - name: factId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of external links
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LinkListResponse"
  /facts/{factId}/links/{linkId}:
    delete:
      operationId: deleteLink
      tags: [links]
      summary: Delete an external link
      parameters:
        - name: factId
          in: path
          required: true
          schema:
            type: integer
        - name: linkId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Link deleted
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
...
components:
  securitySchemes:
    ApiKey:
      type: apiKey
      in: header
      name: X-API-Key
      description: Static API key set via the ADMIN_API_KEY environment variable. Used by automated callers (e.g. LLM agents) that cannot use a browser session.
    SessionCookie:
      type: apiKey
      in: cookie
      name: sid
      description: Browser session cookie (set by the login flow).
...
  schemas:
...
    DuplicateConflict:
      type: object
      required: [error, isDuplicate, confidence]
      properties:
        error:
          type: string
        isDuplicate:
          type: boolean
        confidence:
          type: integer
        matchingFactId:
          type: integer
        matchingFactText:
          type: string
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    FactSummary:
      type: object
      required: [id, text, upvotes, downvotes, commentCount, hashtags, createdAt]
      properties:
        id:
          type: integer
        text:
          type: string
        upvotes:
          type: integer
        downvotes:
          type: integer
        score:
          type: integer
        commentCount:
          type: integer
        hashtags:
          type: array
          items:
            type: string
        submittedBy:
          type: string
          nullable: true
        submittedByImage:
          type: string
          nullable: true
        userRating:
          type: string
          enum: [up, down]
          nullable: true
        useCase:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
    FactDetail:
      allOf:
        - $ref: "#/components/schemas/FactSummary"
        - type: object
          properties:
            links:
              type: array
              items:
                $ref: "#/components/schemas/ExternalLink"
            parentId:
              type: integer
              nullable: true
            variants:
              type: array
              items:
                $ref: "#/components/schemas/FactSummary"
    FactListResponse:
      type: object
      required: [facts, total]
      properties:
        facts:
          type: array
          items:
            $ref: "#/components/schemas/FactSummary"
        total:
          type: integer
    CreateFactRequest:
      type: object
      required: [text, captchaToken]
      properties:
        text:
          type: string
          minLength: 10
          maxLength: 1000
        hashtags:
          type: array
          items:
            type: string
          default: []
        captchaToken:
          type: string
        skipDuplicateCheck:
          type: boolean
          description: If true, bypasses server-side duplicate enforcement and allows submission even when a duplicate is detected.
    RateFactRequest:
      type: object
      required: [rating]
      properties:
        rating:
          type: string
          enum: [up, down, none]
    RatingResult:
      type: object
      required: [upvotes, downvotes, userRating]
      properties:
        upvotes:
          type: integer
        downvotes:
          type: integer
        userRating:
          type: string
          enum: [up, down]
          nullable: true
    Comment:
      type: object
      required: [id, factId, text, createdAt]
      properties:
        id:
          type: integer
        factId:
          type: integer
        text:
          type: string
        authorId:
          type: string
          nullable: true
        authorName:
          type: string
          nullable: true
        authorImage:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
    CommentListResponse:
      type: object
      required: [comments, total]
      properties:
        comments:
          type: array
          items:
            $ref: "#/components/schemas/Comment"
        total:
          type: integer
    AddCommentRequest:
      type: object
      required: [text, captchaToken]
      properties:
        text:
          type: string
          minLength: 1
          maxLength: 2000
        captchaToken:
          type: string
    ExternalLink:
      type: object
      required: [id, factId, url, createdAt]
      properties:
        id:
          type: integer
        factId:
          type: integer
        url:
          type: string
        title:
          type: string
          nullable: true
        platform:
          type: string
          nullable: true
        addedBy:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
    LinkListResponse:
      type: object
      required: [links]
      properties:
        links:
          type: array
          items:
            $ref: "#/components/schemas/ExternalLink"
...